+++ sdk/security/keyvault/azsecrets/client.go
@@ -31,7 +31,7 @@ type Client struct {
 // be downloaded. This operation requires the secrets/backup permission.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 7.4
+// Generated from API version 7.5-preview.1
 //   - name - The name of the secret.
 //   - options - BackupSecretOptions contains the optional parameters for the Client.BackupSecret method.
 func (client *Client) BackupSecret(ctx context.Context, name string, options *BackupSecretOptions) (BackupSecretResponse, error) {
@@ -61,7 +61,7 @@ func (client *Client) backupSecretCreateRequest(ctx context.Context, name string
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "7.4")
+	reqQP.Set("api-version", "7.5-preview.1")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -80,7 +80,7 @@ func (client *Client) backupSecretHandleResponse(resp *http.Response) (BackupSec
 // version of a secret. This operation requires the secrets/delete permission.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 7.4
+// Generated from API version 7.5-preview.1
 //   - name - The name of the secret.
 //   - options - DeleteSecretOptions contains the optional parameters for the Client.DeleteSecret method.
 func (client *Client) DeleteSecret(ctx context.Context, name string, options *DeleteSecretOptions) (DeleteSecretResponse, error) {
@@ -110,7 +110,7 @@ func (client *Client) deleteSecretCreateRequest(ctx context.Context, name string
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "7.4")
+	reqQP.Set("api-version", "7.5-preview.1")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -129,7 +129,7 @@ func (client *Client) deleteSecretHandleResponse(resp *http.Response) (DeleteSec
 // operation requires the secrets/get permission.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 7.4
+// Generated from API version 7.5-preview.1
 //   - name - The name of the secret.
 //   - options - GetDeletedSecretOptions contains the optional parameters for the Client.GetDeletedSecret method.
 func (client *Client) GetDeletedSecret(ctx context.Context, name string, options *GetDeletedSecretOptions) (GetDeletedSecretResponse, error) {
@@ -159,7 +159,7 @@ func (client *Client) getDeletedSecretCreateRequest(ctx context.Context, name st
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "7.4")
+	reqQP.Set("api-version", "7.5-preview.1")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -178,7 +178,7 @@ func (client *Client) getDeletedSecretHandleResponse(resp *http.Response) (GetDe
 // permission.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 7.4
+// Generated from API version 7.5-preview.1
 //   - name - The name of the secret.
 //   - version - The version of the secret. This URI fragment is optional. If not specified, the latest version of the secret
 //     is returned.
@@ -211,7 +211,7 @@ func (client *Client) getSecretCreateRequest(ctx context.Context, name string, v
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "7.4")
+	reqQP.Set("api-version", "7.5-preview.1")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -229,7 +229,7 @@ func (client *Client) getSecretHandleResponse(resp *http.Response) (GetSecretRes
 // NewListDeletedSecretPropertiesPager - The Get Deleted Secrets operation returns the secrets that have been deleted for
 // a vault enabled for soft-delete. This operation requires the secrets/list permission.
 //
-// Generated from API version 7.4
+// Generated from API version 7.5-preview.1
 //   - options - ListDeletedSecretPropertiesOptions contains the optional parameters for the Client.NewListDeletedSecretPropertiesPager
 //     method.
 func (client *Client) NewListDeletedSecretPropertiesPager(options *ListDeletedSecretPropertiesOptions) *runtime.Pager[ListDeletedSecretPropertiesResponse] {
@@ -268,7 +268,7 @@ func (client *Client) listDeletedSecretPropertiesCreateRequest(ctx context.Conte
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "7.4")
+	reqQP.Set("api-version", "7.5-preview.1")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -287,7 +287,7 @@ func (client *Client) listDeletedSecretPropertiesHandleResponse(resp *http.Respo
 // identifier and its attributes are provided in the response. Individual secret versions are not listed in the
 // response. This operation requires the secrets/list permission.
 //
-// Generated from API version 7.4
+// Generated from API version 7.5-preview.1
 //   - options - ListSecretPropertiesOptions contains the optional parameters for the Client.NewListSecretPropertiesPager
 //     method.
 func (client *Client) NewListSecretPropertiesPager(options *ListSecretPropertiesOptions) *runtime.Pager[ListSecretPropertiesResponse] {
@@ -326,7 +326,7 @@ func (client *Client) listSecretPropertiesCreateRequest(ctx context.Context, opt
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "7.4")
+	reqQP.Set("api-version", "7.5-preview.1")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -344,7 +344,7 @@ func (client *Client) listSecretPropertiesHandleResponse(resp *http.Response) (L
 // NewListSecretPropertiesVersionsPager - The full secret identifier and attributes are provided in the response. No values
 // are returned for the secrets. This operations requires the secrets/list permission.
 //
-// Generated from API version 7.4
+// Generated from API version 7.5-preview.1
 //   - name - The name of the secret.
 //   - options - ListSecretPropertiesVersionsOptions contains the optional parameters for the Client.NewListSecretPropertiesVersionsPager
 //     method.
@@ -388,7 +388,7 @@ func (client *Client) listSecretPropertiesVersionsCreateRequest(ctx context.Cont
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "7.4")
+	reqQP.Set("api-version", "7.5-preview.1")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -408,7 +408,7 @@ func (client *Client) listSecretPropertiesVersionsHandleResponse(resp *http.Resp
 // secrets/purge permission.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 7.4
+// Generated from API version 7.5-preview.1
 //   - name - The name of the secret.
 //   - options - PurgeDeletedSecretOptions contains the optional parameters for the Client.PurgeDeletedSecret method.
 func (client *Client) PurgeDeletedSecret(ctx context.Context, name string, options *PurgeDeletedSecretOptions) (PurgeDeletedSecretResponse, error) {
@@ -438,7 +438,7 @@ func (client *Client) purgeDeletedSecretCreateRequest(ctx context.Context, name 
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "7.4")
+	reqQP.Set("api-version", "7.5-preview.1")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -448,7 +448,7 @@ func (client *Client) purgeDeletedSecretCreateRequest(ctx context.Context, name 
 // enabled vault. This operation requires the secrets/recover permission.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 7.4
+// Generated from API version 7.5-preview.1
 //   - name - The name of the deleted secret.
 //   - options - RecoverDeletedSecretOptions contains the optional parameters for the Client.RecoverDeletedSecret method.
 func (client *Client) RecoverDeletedSecret(ctx context.Context, name string, options *RecoverDeletedSecretOptions) (RecoverDeletedSecretResponse, error) {
@@ -478,7 +478,7 @@ func (client *Client) recoverDeletedSecretCreateRequest(ctx context.Context, nam
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "7.4")
+	reqQP.Set("api-version", "7.5-preview.1")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -497,7 +497,7 @@ func (client *Client) recoverDeletedSecretHandleResponse(resp *http.Response) (R
 // permission.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 7.4
+// Generated from API version 7.5-preview.1
 //   - parameters - The parameters to restore the secret.
 //   - options - RestoreSecretOptions contains the optional parameters for the Client.RestoreSecret method.
 func (client *Client) RestoreSecret(ctx context.Context, parameters RestoreSecretParameters, options *RestoreSecretOptions) (RestoreSecretResponse, error) {
@@ -523,7 +523,7 @@ func (client *Client) restoreSecretCreateRequest(ctx context.Context, parameters
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "7.4")
+	reqQP.Set("api-version", "7.5-preview.1")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, parameters)
@@ -542,7 +542,7 @@ func (client *Client) restoreSecretHandleResponse(resp *http.Response) (RestoreS
 // creates a new version of that secret. This operation requires the secrets/set permission.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 7.4
+// Generated from API version 7.5-preview.1
 //   - name - The name of the secret. The value you provide may be copied globally for the purpose of running the service.
 //     The value provided should not include personally identifiable or sensitive information.
 //   - parameters - The parameters for setting the secret.
@@ -574,7 +574,7 @@ func (client *Client) setSecretCreateRequest(ctx context.Context, name string, p
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "7.4")
+	reqQP.Set("api-version", "7.5-preview.1")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, parameters)
@@ -594,7 +594,7 @@ func (client *Client) setSecretHandleResponse(resp *http.Response) (SetSecretRes
 // This operation requires the secrets/set permission.
 // If the operation fails it returns an *azcore.ResponseError type.
 //
-// Generated from API version 7.4
+// Generated from API version 7.5-preview.1
 //   - name - The name of the secret.
 //   - version - The version of the secret.
 //   - parameters - The parameters for update secret operation.
@@ -627,7 +627,7 @@ func (client *Client) updateSecretPropertiesCreateRequest(ctx context.Context, n
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "7.4")
+	reqQP.Set("api-version", "7.5-preview.1")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, runtime.MarshalAsJSON(req, parameters)